# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the workdadventure back office written at the Hochschule Bochum",
        "title": "Workadventure Admin Back Office API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/bo-mathadventure/admin"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://github.com/teamdigitale/licenses/blob/master/AGPL-3.0-or-later"
        },
        "version": "1.0"
    },
    "host": "localhost:4664",
    "basePath": "/",
    "paths": {
...
        "/system/admin/map": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all maps. Requires permission admin.map.view or admin.map.edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List maps",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/admin.AdminMapResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Requires permission admin.map.edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create map",
                "parameters": [
                    {
                        "description": "-",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateMap"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminMapResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/system/admin/map/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get map by ID. Requires permission admin.map.view or admin.map.edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get specific map",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Map ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminMapResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update map by ID. Requires permission admin.map.edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update map",
                "parameters": [
                    {
                        "description": "-",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateMap"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Map ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.AdminMapResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete map by ID. Requires permission admin.map.edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete map",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Map ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AdminMapResponse": {
            "type": "object",
            "properties": {
                "canReport": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "object",
                    "properties": {
                        "enable": {
                            "type": "boolean"
                        },
                        "enableDisconnectedList": {
                            "type": "boolean"
                        },
                        "enableOnlineList": {
                            "type": "boolean"
                        },
                        "enableUpload": {
                            "type": "boolean"
                        }
                    }
                },
                "contactPage": {
                    "type": "string",
                    "example": "https://mycompany.com/contact-us"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                },
                "expireOn": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                },
                "id": {
                    "type": "integer"
                },
                "mapUrl": {
                    "type": "string",
                    "example": "/_/global/thecodingmachine.github.io/workadventure-map-starter-kit/map.tmj"
                },
                "policy": {
                    "type": "string",
                    "enum": [
                        "anonymous",
                        "login",
                        "permission"
                    ]
                },
                "roomName": {
                    "type": "string",
                    "example": "Default Room"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "admin",
                        "editor"
                    ]
                }
            }
        },
...
        "admin.CreateMap": {
            "type": "object",
            "properties": {
                "canReport": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "object",
                    "properties": {
                        "enable": {
                            "type": "boolean"
                        },
                        "enableDisconnectedList": {
                            "type": "boolean"
                        },
                        "enableOnlineList": {
                            "type": "boolean"
                        },
                        "enableUpload": {
                            "type": "boolean"
                        }
                    }
                },
                "contactPage": {
                    "type": "string",
                    "example": "https://mycompany.com/contact-us"
                },
                "expireOn": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                },
                "mapUrl": {
                    "type": "string",
                    "example": "/_/global/thecodingmachine.github.io/workadventure-map-starter-kit/map.tmj"
                },
                "policy": {
                    "type": "string",
                    "enum": [
                        "anonymous",
                        "login",
                        "permission"
                    ]
                },
                "roomName": {
                    "type": "string",
                    "example": "Default Room"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "admin",
                        "editor"
                    ]
                }
            }
        },
...
        "admin.UpdateMap": {
            "type": "object",
            "properties": {
                "canReport": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "object",
                    "properties": {
                        "enable": {
                            "type": "boolean"
                        },
                        "enableDisconnectedList": {
                            "type": "boolean"
                        },
                        "enableOnlineList": {
                            "type": "boolean"
                        },
                        "enableUpload": {
                            "type": "boolean"
                        }
                    }
                },
                "contactPage": {
                    "type": "string",
                    "example": "https://mycompany.com/contact-us"
                },
                "expireOn": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z07:00"
                },
                "mapUrl": {
                    "type": "string",
                    "example": "/_/global/thecodingmachine.github.io/workadventure-map-starter-kit/map.tmj"
                },
                "policy": {
                    "type": "string",
                    "enum": [
                        "anonymous",
                        "login",
                        "permission"
                    ]
                },
                "roomName": {
                    "type": "string",
                    "example": "Default Room"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "admin",
                        "editor"
                    ]
                }
            }
        },
...
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "extra": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "ERR_INVALID_PERMISSIONS"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT user token from login",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}